Add Permission.users_with_perm()
It is now possible to query for the set of users that have a permission
on a program, with the Permission.users_with_perm() classmethod.

To avoid performance costs, Permission.user_has_perm() and
ESPUser.isAdministrator() DO NOT call this function.

Adds more checks in PermissionTestCase, to make sure that
ESPUser.isAdministrator() works as expected.

This implementation might not be performant. Someone should check that
this doesn't degrade performance of Permissions.
(cherry picked from commit ef4479377aea49ae871adcf7efd9577eae439347)